// src/models/Quotes.ts
export interface Quote {
    exchange: string
    shortname: string
    quoteType: string
    symbol: string
    index: string
    score: number
    typeDisp: string
    longname: string
    exchDisp: string
    sector: string
    sectorDisp: string
    industry: string
    industryDisp: string
    dispSecIndFlag: boolean
    logoUrl: string
    isYahooFinance: boolean
}